+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Remember to align the itemized text with the first line of an item within a list
 
 ## jax 0.4.9
 
+* Deprecations
+  * The `in_axis_resources` and `out_axis_resources` arguments of pjit have been
+    deprecated. Please use `in_shardings` and `out_shardings` respectively.
+
 ## jaxlib 0.4.9
 
 ## jax 0.4.8 (March 29, 2023)